+++ src/types/index.ts
@@ -2,6 +2,6 @@
  * Copyright (c) 2024-2026 Cerious DevTech LLC. All rights reserved.
  */
 
-export * from './callback-types';
-export * from './scroll-types';
-export * from './option-types';
+export * from './callback-types.js';
+export * from './scroll-types.js';
+export * from './option-types.js';